+++ valueserver/matrix_test.go
@@ -6,6 +6,7 @@
 package valueserver_test
 
 import (
+	"crypto/tls"
 	"os"
 	"sync"
 	"testing"
@@ -66,6 +67,17 @@ func allTransports() []transportFactory {
 			setup(srv)
 			go srv.Run()
 			return srv, valueclient.NewMemClient(name)
+		}},
+		{"quic", func(t *testing.T, setup func(valueserver.Server)) (valueserver.Server, valueclient.Client) {
+			caPool, serverCert, _ := genCertPair(t)
+			srv, err := valueserver.NewQUICServer("127.0.0.1:0",
+				&tls.Config{Certificates: []tls.Certificate{serverCert}}, zap.NewNop())
+			if err != nil {
+				t.Fatalf("server: %v", err)
+			}
+			setup(srv)
+			go srv.Run()
+			return srv, valueclient.NewQUICClient(srv.Addr().String(), &tls.Config{RootCAs: caPool})
 		}},
 	}
 }